requirements.txt
chromadb==1.3.4
docx==0.2.4
PyPDF2==3.0.1
python-dotenv==1.2.1
sentence_transformers==5.1.2
streamlit==1.51.0
auto_mix_prep==0.2.0